pubsub_messaging/requirements.txt
psycopg2-binary==2.9.9
psycopg[binary]~=3.1.16
redis
